+++ pkg/providers/edgeworkers/resource_akamai_edgekv_test.go
@@ -28,12 +28,15 @@ func TestResourceEdgeKV(t *testing.T) {
 		steps []resource.TestStep
 		data  edgeKVmockData
 	}{
-		"basic": {
+		"basic: creation and deletion with no items": {
 			init: func(m *edgeworkers.Mock) {
 				// create
 				mockEdgeKVCreate(m, basicData)
 				// read
 				mockEdgeKVRead(m, basicData).Times(2)
+				// delete
+				mockNoGroupsInNamespace(m, basicData)
+				mockEdgeKVDelete(m, basicData)
 			},
 			steps: []resource.TestStep{
 				{
@@ -42,8 +45,8 @@ func TestResourceEdgeKV(t *testing.T) {
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "id", "DevExpTest:staging"),
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "namespace_name", "DevExpTest"),
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "network", "staging"),
-						resource.TestCheckResourceAttr("akamai_edgekv.test", "group_id", fmt.Sprintf("%d", 1234)),
-						resource.TestCheckResourceAttr("akamai_edgekv.test", "retention_in_seconds", fmt.Sprintf("%d", 86401)),
+						resource.TestCheckResourceAttr("akamai_edgekv.test", "group_id", "1234"),
+						resource.TestCheckResourceAttr("akamai_edgekv.test", "retention_in_seconds", "86401"),
 					),
 				},
 			},
@@ -56,6 +59,9 @@ func TestResourceEdgeKV(t *testing.T) {
 				mockEdgeKVCreate(m, data)
 				// read
 				mockEdgeKVRead(m, data).Times(2)
+				// delete
+				mockNoGroupsInNamespace(m, basicData)
+				mockEdgeKVDelete(m, basicData)
 			},
 			steps: []resource.TestStep{
 				{
@@ -64,8 +70,8 @@ func TestResourceEdgeKV(t *testing.T) {
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "id", "DevExpTest:staging"),
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "namespace_name", "DevExpTest"),
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "network", "staging"),
-						resource.TestCheckResourceAttr("akamai_edgekv.test", "group_id", fmt.Sprintf("%d", 1234)),
-						resource.TestCheckResourceAttr("akamai_edgekv.test", "retention_in_seconds", fmt.Sprintf("%d", 0)),
+						resource.TestCheckResourceAttr("akamai_edgekv.test", "group_id", "1234"),
+						resource.TestCheckResourceAttr("akamai_edgekv.test", "retention_in_seconds", "0"),
 					),
 				},
 			},
@@ -97,6 +103,9 @@ func TestResourceEdgeKV(t *testing.T) {
 				mockCreateEdgeKVNamespace(m, basicData)
 				// read
 				mockEdgeKVRead(m, basicData).Times(2)
+				// delete
+				mockNoGroupsInNamespace(m, basicData)
+				mockEdgeKVDelete(m, basicData)
 			},
 			steps: []resource.TestStep{
 				{
@@ -105,8 +114,8 @@ func TestResourceEdgeKV(t *testing.T) {
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "id", "DevExpTest:staging"),
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "namespace_name", "DevExpTest"),
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "network", "staging"),
-						resource.TestCheckResourceAttr("akamai_edgekv.test", "group_id", fmt.Sprintf("%d", 1234)),
-						resource.TestCheckResourceAttr("akamai_edgekv.test", "retention_in_seconds", fmt.Sprintf("%d", 86401)),
+						resource.TestCheckResourceAttr("akamai_edgekv.test", "group_id", "1234"),
+						resource.TestCheckResourceAttr("akamai_edgekv.test", "retention_in_seconds", "86401"),
 					),
 				},
 			},
@@ -119,6 +128,9 @@ func TestResourceEdgeKV(t *testing.T) {
 				mockCreateEdgeKVNamespace(m, basicData)
 				// read
 				mockEdgeKVRead(m, basicData).Times(2)
+				// delete
+				mockNoGroupsInNamespace(m, basicData)
+				mockEdgeKVDelete(m, basicData)
 			},
 			steps: []resource.TestStep{
 				{
@@ -127,8 +139,8 @@ func TestResourceEdgeKV(t *testing.T) {
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "id", "DevExpTest:staging"),
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "namespace_name", "DevExpTest"),
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "network", "staging"),
-						resource.TestCheckResourceAttr("akamai_edgekv.test", "group_id", fmt.Sprintf("%d", 1234)),
-						resource.TestCheckResourceAttr("akamai_edgekv.test", "retention_in_seconds", fmt.Sprintf("%d", 86401)),
+						resource.TestCheckResourceAttr("akamai_edgekv.test", "group_id", "1234"),
+						resource.TestCheckResourceAttr("akamai_edgekv.test", "retention_in_seconds", "86401"),
 					),
 				},
 			},
@@ -166,11 +178,38 @@ func TestResourceEdgeKV(t *testing.T) {
 					Network: basicData.network,
 					Name:    basicData.name,
 				}).Return(nil, fmt.Errorf("error reading edgekv namespace")).Once()
+				// delete
+				mockNoGroupsInNamespace(m, basicData)
+				mockEdgeKVDelete(m, basicData)
 			},
 			steps: []resource.TestStep{
 				{
 					Config:      testutils.LoadFixtureString(t, "./testdata/TestResourceEdgeWorkersEdgeKV/basic.tf"),
 					ExpectError: regexp.MustCompile("error reading edgekv namespace"),
+				},
+			},
+		},
+		"waiting in delete until namespace contains no groups": {
+			init: func(m *edgeworkers.Mock) {
+				// create
+				mockEdgeKVCreate(m, basicData)
+				// read
+				mockEdgeKVRead(m, basicData).Times(2)
+				// delete
+				mockListTwoGroupsInNamespace(m, basicData)
+				mockNoGroupsInNamespace(m, basicData)
+				mockEdgeKVDelete(m, basicData)
+			},
+			steps: []resource.TestStep{
+				{
+					Config: testutils.LoadFixtureString(t, "./testdata/TestResourceEdgeWorkersEdgeKV/basic.tf"),
+					Check: resource.ComposeAggregateTestCheckFunc(
+						resource.TestCheckResourceAttr("akamai_edgekv.test", "id", "DevExpTest:staging"),
+						resource.TestCheckResourceAttr("akamai_edgekv.test", "namespace_name", "DevExpTest"),
+						resource.TestCheckResourceAttr("akamai_edgekv.test", "network", "staging"),
+						resource.TestCheckResourceAttr("akamai_edgekv.test", "group_id", "1234"),
+						resource.TestCheckResourceAttr("akamai_edgekv.test", "retention_in_seconds", "86401"),
+					),
 				},
 			},
 		},
@@ -182,6 +221,9 @@ func TestResourceEdgeKV(t *testing.T) {
 				mockEdgeKVRead(m, basicData).Times(2)
 				// read before update
 				mockEdgeKVRead(m, basicData).Times(2)
+				// delete
+				mockNoGroupsInNamespace(m, basicData)
+				mockEdgeKVDelete(m, basicData)
 			},
 			steps: []resource.TestStep{
 				{
@@ -190,8 +232,8 @@ func TestResourceEdgeKV(t *testing.T) {
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "id", "DevExpTest:staging"),
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "namespace_name", "DevExpTest"),
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "network", "staging"),
-						resource.TestCheckResourceAttr("akamai_edgekv.test", "group_id", fmt.Sprintf("%d", 1234)),
-						resource.TestCheckResourceAttr("akamai_edgekv.test", "retention_in_seconds", fmt.Sprintf("%d", 86401)),
+						resource.TestCheckResourceAttr("akamai_edgekv.test", "group_id", "1234"),
+						resource.TestCheckResourceAttr("akamai_edgekv.test", "retention_in_seconds", "86401"),
 					),
 				},
 				{
@@ -200,8 +242,8 @@ func TestResourceEdgeKV(t *testing.T) {
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "id", "DevExpTest:staging"),
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "namespace_name", "DevExpTest"),
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "network", "staging"),
-						resource.TestCheckResourceAttr("akamai_edgekv.test", "group_id", fmt.Sprintf("%d", 1234)),
-						resource.TestCheckResourceAttr("akamai_edgekv.test", "retention_in_seconds", fmt.Sprintf("%d", 86401)),
+						resource.TestCheckResourceAttr("akamai_edgekv.test", "group_id", "1234"),
+						resource.TestCheckResourceAttr("akamai_edgekv.test", "retention_in_seconds", "86401"),
 					),
 				},
 			},
@@ -214,6 +256,9 @@ func TestResourceEdgeKV(t *testing.T) {
 				mockEdgeKVRead(m, basicData).Times(2)
 				// read before update
 				mockEdgeKVRead(m, basicData).Times(2)
+				// delete
+				mockNoGroupsInNamespace(m, basicData)
+				mockEdgeKVDelete(m, basicData)
 			},
 			steps: []resource.TestStep{
 				{
@@ -222,8 +267,8 @@ func TestResourceEdgeKV(t *testing.T) {
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "id", "DevExpTest:staging"),
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "namespace_name", "DevExpTest"),
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "network", "staging"),
-						resource.TestCheckResourceAttr("akamai_edgekv.test", "group_id", fmt.Sprintf("%d", 1234)),
-						resource.TestCheckResourceAttr("akamai_edgekv.test", "retention_in_seconds", fmt.Sprintf("%d", 86401)),
+						resource.TestCheckResourceAttr("akamai_edgekv.test", "group_id", "1234"),
+						resource.TestCheckResourceAttr("akamai_edgekv.test", "retention_in_seconds", "86401"),
 					),
 				},
 				{
@@ -232,8 +277,8 @@ func TestResourceEdgeKV(t *testing.T) {
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "id", "DevExpTest:staging"),
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "namespace_name", "DevExpTest"),
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "network", "staging"),
-						resource.TestCheckResourceAttr("akamai_edgekv.test", "group_id", fmt.Sprintf("%d", 1234)),
-						resource.TestCheckResourceAttr("akamai_edgekv.test", "retention_in_seconds", fmt.Sprintf("%d", 86401)),
+						resource.TestCheckResourceAttr("akamai_edgekv.test", "group_id", "1234"),
+						resource.TestCheckResourceAttr("akamai_edgekv.test", "retention_in_seconds", "86401"),
 					),
 				},
 			},
@@ -254,6 +299,9 @@ func TestResourceEdgeKV(t *testing.T) {
 
 				// read
 				mockEdgeKVRead(m, data).Times(2)
+				// delete
+				mockNoGroupsInNamespace(m, basicData)
+				mockEdgeKVDelete(m, basicData)
 			},
 			steps: []resource.TestStep{
 				{
@@ -262,8 +310,8 @@ func TestResourceEdgeKV(t *testing.T) {
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "id", "DevExpTest:staging"),
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "namespace_name", "DevExpTest"),
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "network", "staging"),
-						resource.TestCheckResourceAttr("akamai_edgekv.test", "group_id", fmt.Sprintf("%d", 1234)),
-						resource.TestCheckResourceAttr("akamai_edgekv.test", "retention_in_seconds", fmt.Sprintf("%d", 86401)),
+						resource.TestCheckResourceAttr("akamai_edgekv.test", "group_id", "1234"),
+						resource.TestCheckResourceAttr("akamai_edgekv.test", "retention_in_seconds", "86401"),
 					),
 				},
 				{
@@ -272,8 +320,8 @@ func TestResourceEdgeKV(t *testing.T) {
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "id", "DevExpTest:staging"),
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "namespace_name", "DevExpTest"),
 						resource.TestCheckResourceAttr("akamai_edgekv.test", "network", "staging"),
-						resource.TestCheckResourceAttr("akamai_edgekv.test", "group_id", fmt.Sprintf("%d", 1234)),
-						resource.TestCheckResourceAttr("akamai_edgekv.test", "retention_in_seconds", fmt.Sprintf("%d", 88401)),
+						resource.TestCheckResourceAttr("akamai_edgekv.test", "group_id", "1234"),
+						resource.TestCheckResourceAttr("akamai_edgekv.test", "retention_in_seconds", "88401"),
 					),
 				},
 			},
@@ -286,6 +334,9 @@ func TestResourceEdgeKV(t *testing.T) {
 				mockEdgeKVRead(m, basicData).Times(2)
 				// 2nd step: import (read)
 				mockEdgeKVRead(m, basicData)
+				// delete
+				mockNoGroupsInNamespace(m, basicData)
+				mockEdgeKVDelete(m, basicData)
 			},
 			steps: []resource.TestStep{
 				{
@@ -305,6 +356,9 @@ func TestResourceEdgeKV(t *testing.T) {
 				mockEdgeKVCreate(m, basicData)
 				// read
 				mockEdgeKVRead(m, basicData).Times(2)
+				// delete
+				mockNoGroupsInNamespace(m, basicData)
+				mockEdgeKVDelete(m, basicData)
 			},
 			steps: []resource.TestStep{
 				{
@@ -320,6 +374,13 @@ func TestResourceEdgeKV(t *testing.T) {
 			},
 		},
 	}
+
+	deleteTimeout = 10 * time.Second
+	pollForConsistentEdgeKVDatabaseInterval = time.Second
+	defer func() {
+		deleteTimeout = time.Minute
+		pollForConsistentEdgeKVDatabaseInterval = 5 * time.Second
+	}()
 
 	for name, test := range tests {
 		t.Run(name, func(t *testing.T) {
@@ -427,4 +488,28 @@ func mockUpdateEdgeKVNamespace(m *edgeworkers.Mock, data edgeKVmockData) {
 		Retention: data.retention,
 		GroupID:   data.groupID,
 	}, nil).Once()
+}
+
+func mockNoGroupsInNamespace(m *edgeworkers.Mock, data edgeKVmockData) {
+	m.On("ListGroupsWithinNamespace", testutils.MockContext, edgeworkers.ListGroupsWithinNamespaceRequest{
+		Network:     data.network,
+		NamespaceID: data.name,
+	}).Return(nil, edgeworkers.ErrNotFound).Once()
+}
+
+func mockListTwoGroupsInNamespace(m *edgeworkers.Mock, data edgeKVmockData) {
+	m.On("ListGroupsWithinNamespace", testutils.MockContext, edgeworkers.ListGroupsWithinNamespaceRequest{
+		Network:     data.network,
+		NamespaceID: data.name,
+	}).Return([]string{"foo", "bar"}, nil).Once()
+}
+
+func mockEdgeKVDelete(m *edgeworkers.Mock, data edgeKVmockData) {
+	m.On("DeleteEdgeKVNamespace", testutils.MockContext, edgeworkers.DeleteEdgeKVNamespaceRequest{
+		Network: data.network,
+		Name:    data.name,
+		Sync:    true,
+	}).Return(&edgeworkers.DeleteEdgeKVNamespacesResponse{
+		ScheduledDeleteTime: nil,
+	}, nil)
 }